\name{bitFlip}
\alias{bitFlip}
\title{Binary Flip (Not) Operator}
\description{
  ...
}
\usage{
bitFlip(a)
}
\arguments{
  \item{a}{binary integer vector.}
}
\value{
  binary integer vector of the same length as \code{a}.
}
%\references{ ~put references to the literature/web site here ~ }
%\author{ ~~who you are~~ }
%\seealso{ ~~objects to See Also as \code{\link{~~fun~~}}, ~~~ }
\examples{
}
\keyword{arith}
\keyword{utilities}
